refactor: 패키지구조를 같은 도메인끼리 있을수 있게 수정하라
