+++ src/main/java/nextstep/subway/dto/SectionRequest.java
@@ -1,4 +1,4 @@
-package nextstep.subway.section.dto;
+package nextstep.subway.dto;
 
 public class SectionRequest {
