+++ src/test/java/nextstep/subway/section/SectionAcceptanceTest.java
@@ -17,9 +17,9 @@ import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
 import java.util.stream.Stream;
 import nextstep.subway.AcceptanceTest;
-import nextstep.subway.line.dto.LineResponse;
+import nextstep.subway.dto.LineResponse;
 import nextstep.subway.global.exception.ExceptionType;
-import nextstep.subway.section.dto.SectionRequest;
+import nextstep.subway.dto.SectionRequest;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
